scrolling is now more accurate at the cost of obvious issues with horizontal scrolling
tiles at the edges of nametables get messed up and some tiles' palettes aren't right when partially scrolled

+++ src/ppu.h
@@ -49,7 +49,7 @@ typedef struct {
 	uint16_t vramAddr;
 	uint8_t scrollX;
 	uint8_t scrollY;
-	uint8_t t;
+	uint16_t t;
 	uint8_t x;
 	uint8_t w;
 	uint8_t mirror;
